+++ src/rules/string-quotes/__tests__/index.js
@@ -83,9 +83,16 @@ const scssTestRule = ruleTester(rule, ruleName, {
 
 scssTestRule("double", tr => {
   tr.ok("a {\n  // 'horse'\n}", "ignores single-line SCSS comment")
+
   tr.notOk("a::before {\n  // 'horse'\n  content: 'thing'; }", {
     message: messages.expected("double"),
     line: 3,
-    column: 10,
+    column: 12,
   }, "pays attention when single-line SCSS comment ends")
+
+  tr.notOk("a::before {\n// one\n// two\n// three\n  content: 'thing'; }", {
+    message: messages.expected("double"),
+    line: 5,
+    column: 12,
+  }, "accurate position after // comments")
 })